+++ src/Contract.ts
@@ -5,12 +5,11 @@ import { Serializer } from "./Serializer";
 import {
   TransactionJsonWait,
   Abi,
-  TransactionOptions,
+  CallContractOptions,
   DecodedOperationJson,
   OperationJson,
   DeployOptions,
   TransactionReceipt,
-  SendTransactionOptions,
 } from "./interface";
 import { decodeBase58, encodeBase58, encodeBase64url } from "./utils";
 
@@ -75,7 +74,7 @@ export class Contract {
   /**
    * Contract ID
    */
-  id?: Uint8Array;
+  id: Uint8Array;
 
   /**
    * Set of functions to interact with the smart
@@ -110,7 +109,7 @@ export class Contract {
   functions: {
     [x: string]: <T = Record<string, unknown>>(
       args?: unknown,
-      opts?: TransactionOptions
+      opts?: CallContractOptions
     ) => Promise<{
       operation: OperationJson;
       transaction?: TransactionJsonWait;
@@ -149,13 +148,13 @@ export class Contract {
    * By default it set rc_limit to 1e8, sendTransaction true,
    * sendAbis true, and nonce undefined (to get it from the blockchain)
    */
-  options: TransactionOptions;
+  options: CallContractOptions;
 
   constructor(c: {
     id?: string;
     abi?: Abi;
     bytecode?: Uint8Array;
-    options?: TransactionOptions;
+    options?: CallContractOptions;
     signer?: Signer;
     provider?: Provider;
     /**
@@ -166,8 +165,13 @@ export class Contract {
      */
     serializer?: Serializer;
   }) {
-    if (c.id) this.id = decodeBase58(c.id);
     this.signer = c.signer;
+    if (c.id) this.id = decodeBase58(c.id);
+    else {
+      if (!this.signer)
+        throw new Error("at least signer or contract id must be defined");
+      this.id = decodeBase58(this.signer.getAddress());
+    }
     this.provider = c.provider || c.signer?.provider;
     this.abi = c.abi;
     this.bytecode = c.bytecode;
@@ -189,7 +193,7 @@ export class Contract {
       Object.keys(this.abi.methods).forEach((name) => {
         this.functions[name] = async <T = Record<string, unknown>>(
           argu: unknown = {},
-          options?: TransactionOptions
+          options?: CallContractOptions
         ): Promise<{
           operation: OperationJson;
           transaction?: TransactionJsonWait;
@@ -201,7 +205,7 @@ export class Contract {
             throw new Error("Methods are not defined");
           if (!this.abi.methods[name])
             throw new Error(`Method ${name} not defined in the ABI`);
-          const opts: TransactionOptions = {
+          const opts: CallContractOptions = {
             ...this.options,
             ...options,
           };
@@ -262,15 +266,10 @@ export class Contract {
             ],
           });
 
-          const optsSend: SendTransactionOptions = {
-            broadcast: opts.broadcast,
-            beforeSend: opts.beforeSend,
-          };
-
           if (opts.sendAbis) {
-            optsSend.abis = {};
-            const contractId = encodeBase58(this.id as Uint8Array);
-            optsSend.abis[contractId] = this.abi;
+            if (!opts.abis) opts.abis = {};
+            const contractId = this.getId();
+            if (!opts.abis[contractId]) opts.abis[contractId] = this.abi;
           }
 
           // return result if the transaction will not be broadcasted
@@ -279,13 +278,16 @@ export class Contract {
               throw new Error("This transaction was not broadcasted");
             };
             if (opts.signTransaction)
-              tx = await this.signer.signTransaction(tx, optsSend.abis);
+              tx = await this.signer.signTransaction(
+                tx,
+                opts.sendAbis ? opts.abis : undefined
+              );
             return { operation, transaction: { ...tx, wait: noWait } };
           }
 
           const { transaction, receipt } = await this.signer.sendTransaction(
             tx,
-            optsSend
+            opts
           );
           return { operation, transaction, receipt };
         };
@@ -297,7 +299,6 @@ export class Contract {
    * Get contract Id
    */
   getId(): string {
-    if (!this.id) throw new Error("id is not defined");
     return encodeBase58(this.id);
   }
 
@@ -357,9 +358,7 @@ export class Contract {
       ...options,
     };
 
-    const contractId = this.id
-      ? encodeBase58(this.id)
-      : this.signer.getAddress();
+    const contractId = this.getId();
 
     const operation = {
       upload_contract: {
@@ -398,23 +397,22 @@ export class Contract {
       ],
     });
 
-    const optsSend: SendTransactionOptions = {
-      broadcast: opts.broadcast,
-      beforeSend: opts.beforeSend,
-    };
-
     // return result if the transaction will not be broadcasted
     if (!opts.sendTransaction) {
       const noWait = () => {
         throw new Error("This transaction was not broadcasted");
       };
-      if (opts.signTransaction) tx = await this.signer.signTransaction(tx);
+      if (opts.signTransaction)
+        tx = await this.signer.signTransaction(
+          tx,
+          opts.sendAbis ? opts.abis : undefined
+        );
       return { operation, transaction: { ...tx, wait: noWait } };
     }
 
     const { transaction, receipt } = await this.signer.sendTransaction(
       tx,
-      optsSend
+      opts
     );
     return { operation, transaction, receipt };
   }
@@ -449,7 +447,6 @@ export class Contract {
     if (!this.abi || !this.abi.methods || !this.abi.methods[op.name])
       throw new Error(`Operation ${op.name} unknown`);
     if (!this.serializer) throw new Error("Serializer is not defined");
-    if (!this.id) throw new Error("Contract id is not defined");
     const method = this.abi.methods[op.name];
 
     let bufferArguments = new Uint8Array(0);
@@ -464,7 +461,7 @@ export class Contract {
 
     return {
       call_contract: {
-        contract_id: encodeBase58(this.id),
+        contract_id: this.getId(),
         entry_point: method.entry_point,
         args: encodeBase64url(bufferArguments),
       },
@@ -494,15 +491,14 @@ export class Contract {
    * ```
    */
   async decodeOperation(op: OperationJson): Promise<DecodedOperationJson> {
-    if (!this.id) throw new Error("Contract id is not defined");
     if (!this.abi || !this.abi.methods)
       throw new Error("Methods are not defined");
     if (!this.serializer) throw new Error("Serializer is not defined");
     if (!op.call_contract)
       throw new Error("Operation is not CallContractOperation");
-    if (op.call_contract.contract_id !== encodeBase58(this.id))
+    if (op.call_contract.contract_id !== this.getId())
       throw new Error(
-        `Invalid contract id. Expected: ${encodeBase58(this.id)}. Received: ${
+        `Invalid contract id. Expected: ${this.getId()}. Received: ${
           op.call_contract.contract_id
         }`
       );