+++ src/index.ts
@@ -5,4 +5,5 @@ export * as interfaces from "./interface";
 export * from "./Contract";
 export * from "./Signer";
 export * from "./Provider";
+export * from "./Transaction";
 export * from "./Serializer";